--- source/library/reflect/private/CMakeLists.txt
@@ -1,3 +1,0 @@
-target_sources(reflect PRIVATE
-    ${CMAKE_CURRENT_SOURCE_DIR}/reflect.cpp
-)
