+++ source/library/reflex/public/potato/reflex/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(reflect PRIVATE
+target_sources(reflex PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/_export.h
     ${CMAKE_CURRENT_SOURCE_DIR}/_wrapper.h
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata.h
